+++ jormungandr/src/rest/mod.rs
@@ -1,10 +1,10 @@
 //! REST API of the node
+#[cfg(feature = "evm")]
+mod evm;
 #[cfg(feature = "prometheus-metrics")]
 mod prometheus;
 pub mod v0;
 mod v1;
-#[cfg(feature = "evm")]
-mod v2;
 
 use crate::context::{Context, ContextLock, ServerStopper};
 use futures::{channel::mpsc, prelude::*};
@@ -29,7 +29,7 @@ pub async fn start_rest_server(config: Config, context: ContextLock) {
         .set_rest_server_stopper(ServerStopper::new(stopper_tx));
     let api = v0::filter(context.clone()).or(v1::filter(context.clone()));
     #[cfg(feature = "evm")]
-    let api = api.or(v2::filter(context.clone()));
+    let api = api.or(evm::filter(context.clone()));
 
     let api = warp::path!("api" / ..)
         .and(api)
